moved dropdown subcomponent code into details page
